Separated 2 configuration services conf.resolve and conf.noenum
conf.resolve holds the list of fields for which resolution (DNS, etc) should be done
conf.noenum  holds the list of fields for conversion to string  should NOT be done

Conversions that use the conf.resolve service are off by default. This service
is done for expensive conversions like DNS.
Conversions that use conf.noenum service are on by default. This is mainly done
for Enum fields for which "resolution" is only a matter of a lookup into
a dictionnary